% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{check_csv_file}
\alias{check_csv_file}
\title{Check if the csv file has a correct format}
\usage{
check_csv_file(file)
}
\arguments{
\item{file}{The file's path}
}
\value{
A logical
}
\description{
Check if the csv file has a correct format
}
\examples{
library(dataesgobr)
\dontrun{
correct <- check_csv_file("fichero.csv")
}
}
