% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{check_file}
\alias{check_file}
\title{Check if the dataset has a correct format}
\usage{
check_file(file)
}
\arguments{
\item{file}{The file to check}
}
\value{
Return a logical, if the file is correct it will be TRUE, else FALSE
}
\description{
Check if the dataset has a correct format
}
