% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{confirm_action}
\alias{confirm_action}
\title{Generates an interface in R console with a question and options to answer it}
\usage{
confirm_action(question, options = NULL)
}
\arguments{
\item{question}{character containing the question}

\item{options}{vector containing the options to answer the question}
}
\value{
logical
}
\description{
Generates an interface in R console with a question and options to answer it
}
