% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats-data.R
\docType{data}
\name{datos}
\alias{datos}
\title{Format data on dataesgobr}
\format{A data.frame}
\usage{
data(formats)
}
\description{
The dataset formats contains the extension and formats that dataesgobr
can uses
}
\examples{
formats <- data(formats)
formats
}
\keyword{datasets}
