% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{download_data}
\alias{download_data}
\title{Download files asociate with dataesgobr object}
\usage{
download_data(x, format, all = TRUE, position = 0, noconfirm = FALSE,
  path = NULL)
}
\arguments{
\item{x}{dataesgobr containing information and data from datos.gob.es}

\item{format}{The data's format to download}

\item{all}{This parameter indicates if the function must download every file}

\item{position}{The number in the format list}

\item{noconfirm}{logical This parameter indicates if the user must confirm
the downloads or no}

\item{path}{character The path where the file will be save}
}
\description{
This function downloads the data associated with the dataset
passed like param from datos.gob.es
}
\examples{
library(dataesgobr)
\dontrun{
dataset <- search_by_id('l01350167-atestados-policia-local-de-las-palmas-de-gran-canaria-2013')
download_data(dataset, "text/csv", noconfirm = TRUE)
download_data(dataset, "text/csv", FALSE, 3, noconfirm = TRUE)
download_data(dataset, "application/vnd.oasis.opendocument.spreadsheet", TRUE, noconfirm = TRUE)
download_data(dataset, "application/vnd.oasis.opendocument.spreadsheet", FALSE, position = 6, noconfirm = TRUE)
}
}
