% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{filter_by_description}
\alias{filter_by_description}
\title{Filter data.frame by description using q param}
\usage{
filter_by_description(data, q, quiet = FALSE)
}
\arguments{
\item{data}{A data.frame that will be filtered}

\item{q}{A string to match in data.frame data}

\item{quiet}{Set if the function can print info messages or not}
}
\value{
A data.frame with rows that matches the description
}
\description{
Filter data.frame by description using q param
}
\examples{
library(dataesgobr)
\donttest{
datasets <- search_by_title('salud')
datasetsFiltered <- filter_by_description(datasets, 'salud')
}
}
