% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{filter_by_keywords}
\alias{filter_by_keywords}
\title{Filter data.frame by keywords using keywords param}
\usage{
filter_by_keywords(data, keywords, quiet = FALSE)
}
\arguments{
\item{data}{A data.frame that will be filtered}

\item{keywords}{A string to match in data.frame data}

\item{quiet}{Set if the function can print info messages or not}
}
\value{
A data.frame that matches any given keyword
}
\description{
Filter data.frame by keywords using keywords param
}
\examples{
library(dataesgobr)
\donttest{
datasets <- search_by_title('salud')
datasetsFiltered <- filter_by_keywords(datasets, 'enfermedad')
}
}
