% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{filter_by_title}
\alias{filter_by_title}
\title{Filter data.frame by title using q parameter}
\usage{
filter_by_title(data, q)
}
\arguments{
\item{data}{A data.frame that will be filter}

\item{q}{A string to mach in data.frame data}
}
\value{
A data.frame with rows that match the title
}
\description{
Filter data.frame by title using q parameter
}
