% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{get_all_keywords}
\alias{get_all_keywords}
\title{Obtain keywords in data.frame with the number of ocurrences}
\usage{
get_all_keywords(data)
}
\arguments{
\item{data}{A data.frame with datasets from datos.gob.es}
}
\value{
A list with keywords
}
\description{
Obtain keywords in data.frame with the number of ocurrences
}
\examples{
library(dataesgobr)
\donttest{
datasets <- search_by_title('consumo')
keywords <- get_all_keywords(datasets)
}
}
