% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{get_available_formats}
\alias{get_available_formats}
\title{Generate a data.frame that contains the type of elements,
information and repetitions for each one}
\usage{
get_available_formats(data)
}
\arguments{
\item{data}{A dataesgobr object}
}
\value{
A data.frame
}
\description{
Generate a data.frame that contains the type of elements,
information and repetitions for each one
}
\examples{
library(dataesgobr)
\donttest{
dataesgobrObject <- search_by_id('l01280066-horarios-de-metro1')
formats <- get_available_formats(dataesgobrObject)
formats
}
}
