% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{get_extension}
\alias{get_extension}
\title{Get the extension that matches with the format passed like parameter}
\usage{
get_extension(format)
}
\arguments{
\item{format}{A string that contains the format}
}
\value{
A string
}
\description{
Get the extension that matches with the format passed like parameter
}
\examples{
library(dataesgobr)
extension <- get_extension("text/csv")
extension <- get_extension("application/pdf")
}
