% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{get_formats}
\alias{get_formats}
\title{Generate a list that contains the type of elements and repetitions for
each one}
\usage{
get_formats(data)
}
\arguments{
\item{data}{A dataesgobr object}
}
\value{
A list
}
\description{
Generate a list that contains the type of elements and repetitions for
each one
}
\examples{
library(dataesgobr)
\donttest{
dataesgobrObject <- search_by_id('l01350167-atestados-policia-local-de-las-palmas-de-gran-canaria-2013')
formats <- get_formats(dataesgobrObject)
formats
}
}
