% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{get_name}
\alias{get_name}
\title{Extract the name of the file in the URL}
\usage{
get_name(url, format)
}
\arguments{
\item{url}{A string with URL and the name of the file}

\item{format}{The data's format to check the extension}
}
\value{
A string with the file's name
}
\description{
Extract the name of the file in the URL
}
\examples{
library(dataesgobr)
name <- get_name("archivo-de-datos.csv&rnd=1234", "text/csv")
name <- get_name("https://datos.ayto-arganda.es/contratos.pdf", "application/pdf")
}
