% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{get_publisher}
\alias{get_publisher}
\title{Get the publisher of the dataset}
\usage{
get_publisher(id)
}
\arguments{
\item{id}{A string with the dataset's ID}
}
\value{
A data.frame
}
\description{
Get the publisher of the dataset
}
\examples{
library(dataesgobr)
\donttest{
publisher <- get_publisher("L01280066")
}
}
