% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{get_spatials_from_api}
\alias{get_spatials_from_api}
\title{This method send a request to datos.gob.es and return a data.frame
that contains the spatials available in the catalog}
\usage{
get_spatials_from_api()
}
\value{
A data.frame
}
\description{
This method send a request to datos.gob.es and return a data.frame
that contains the spatials available in the catalog
}
