% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{get_symbol}
\alias{get_symbol}
\title{This function detects the delim from a csv file}
\usage{
get_symbol(file)
}
\arguments{
\item{file}{The file with the content to check}
}
\value{
The symbol as character that split the columns
}
\description{
This function detects the delim from a csv file
}
\examples{
library(dataesgobr)
file <- system.file("extdata", "fichero.csv", package="dataesgobr")
symbol <- get_symbol(file)
}
