% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{make_url}
\alias{make_url}
\title{Loads the file url_params and make urls to extract information from the API}
\usage{
make_url(field, request, params = NULL)
}
\arguments{
\item{field}{String to specifies the field that we want to search}

\item{request}{String with the information to search}

\item{params}{Optional list with the parameters to add to the query}
}
\value{
A string containing the url generated
}
\description{
Loads the file url_params and make urls to extract information from the API
}
\examples{
library(dataesgobr)

# Generates the url associated to the id passed
url <- make_url("id", "a13002908-residentes-en-la-comunidad-de-madrid-por-lugar-de-nacimiento")

# Generates the url associated to the title and the parameters passed
url <- make_url("title", "atestados", c("sort" = "title", "pagesize" = 50, "page" = 1))
}
