% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GUI.R
\name{runGUI}
\alias{runGUI}
\title{Launches the web-based GUI for dataesgobr}
\usage{
runGUI()
}
\value{
nothing
}
\description{
Download, load and edit using an interactive user interface in the web browser
}
\details{
The \pkg{dataesgobr} package contains a basic Shiny-based GUI that you can
use in order to work easily while you are using the package.
You have to install the \pkg{shiny} package to use this GUI.
}
\examples{
\dontrun{
library(dataesgobr)
runGUI()
}
}
