% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{search_by}
\alias{search_by}
\title{Creates a data.frame containing datasets from datos.gob.es that matches
the parameters passed}
\usage{
search_by(title, theme, publisher)
}
\arguments{
\item{title}{Title field}

\item{theme}{Theme field}

\item{publisher}{Publisher field}
}
\value{
A data.frame
}
\description{
Creates a data.frame containing datasets from datos.gob.es that matches
the parameters passed
}
\examples{
library(dataesgobr)
\donttest{
datasets <- search_by('atestados', 'salud', 'L01280066')
}
}
