% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{search_by_id}
\alias{search_by_id}
\title{Creates an object of the class dataesgobr that matches with the
dataset id}
\usage{
search_by_id(id)
}
\arguments{
\item{id}{A string}
}
\value{
A dataesgobr object
}
\description{
Creates an object of the class dataesgobr that matches with the
dataset id
}
\examples{
library(dataesgobr)
\dontrun{
dataesgobr_search_by_id('l01280066-presupuestos-20141')
}
}
