% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{search_by_title}
\alias{search_by_title}
\title{Creates a data.frame containing datasets from datos.gob.es}
\usage{
search_by_title(title, numentry = 50)
}
\arguments{
\item{title}{Title to search}

\item{numentry}{Numbers of datasets}
}
\value{
A data.frame containing information about datasets that match with
title param
}
\description{
Send a request to datos.gob.es using title variable
to search datasets that match with the title, then the results are return
as data.frame
}
\examples{
library(dataesgobr)
\dontrun{
# Return first 50 matches that contain puente in their title
mydataesgobr <- dataesgobr_search_by_title('puente')

# Return the first 78 matches that contain puente in their title and
mydataesgobr <- dataesgobr_search_by_title('puente', 78)

}
}
