% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{check_file}
\alias{check_file}
\title{Check if the dataset has a correct format}
\usage{
check_file(file, path = NULL)
}
\arguments{
\item{file}{The file to check}

\item{path}{character The path of the file, you need to use this parameter if
the file is located a directory different to working directory}
}
\value{
Return a logical, if the file is correct it will be TRUE, else FALSE
}
\description{
Check if the dataset has a correct format
}
\examples{
library(dataesgobr)
file <- system.file("extdata", "fichero.csv", package="dataesgobr")
correct <- check_file(file)
}
