% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{download_data}
\alias{download_data}
\title{Download files asociate with dataesgobr object}
\usage{
download_data(x, format, all = TRUE, position = 0)
}
\arguments{
\item{x}{dataesgobr containing information and data from datos.gob.es}

\item{format}{The data's format to download}
}
\description{
This function downloads the data associated with the dataset
passed like param from datos.gob.es
}
