% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{filter_by_keywords}
\alias{filter_by_keywords}
\title{Filter data.frame by title using q parameter}
\usage{
filter_by_keywords(data, keywords)
}
\arguments{
\item{data}{A data.frame that will be filter}

\item{keywords}{A string to mach in data.frame data}
}
\value{
A data.frame that match some keyword
}
\description{
Filter data.frame by title using q parameter
}
