% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{get_id}
\alias{get_id}
\title{This method extracts the ID from an URL}
\usage{
get_id(url)
}
\arguments{
\item{url}{A string}
}
\value{
A string that contains the ID
}
\description{
This method extracts the ID from an URL
}
\examples{
library(dataesgobr)
id <- get_id("https://datos.gob.es/es/catalogo/l01280066-tramites-salud1")
}
