% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{load_data}
\alias{load_data}
\title{Load data from a file}
\usage{
load_data(file, filename = NULL, outfile = NULL)
}
\arguments{
\item{file}{A file with data previously downloaded}

\item{path}{character The path of the file, you need to use this parameter if
the file is located a directory different to working directory}
}
\value{
A data.frame
}
\description{
This function loads the data from the file passed like param
}
\examples{
library(dataesgobr)
\dontrun{
file <- system.file("extdata", "fichero.csv", package = "dataesgobr")
load_data(file, filename = "nuevo.csv")
load_data(file, outfile = tempdir())
}
}
