% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{load_data}
\alias{load_data}
\title{Load data from a file}
\usage{
load_data(file, path = NULL)
}
\arguments{
\item{file}{A file with data previously downloaded}

\item{path}{character The path of the file, you need to use this parameter if
the file is located a directory different to working directory}
}
\value{
A data.frame
}
\description{
This function loads the data from the file passed like param
}
\examples{
library(dataesgobr)
\dontrun{
dataset <- search_by_id('l01350167-atestados-policia-local-de-las-palmas-de-gran-canaria-2013')
download_data(dataset, "text/csv", noconfirm = TRUE)
load_data('DB_VEH_2013.csv')
}
}
