% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{load_data}
\alias{load_data}
\title{Load data from a file}
\usage{
load_data(file)
}
\arguments{
\item{x}{A file with data previously downloaded}
}
\value{
A data.frame
}
\description{
This function loads the data from the file passed like param
}
