% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{load_data}
\alias{load_data}
\title{Load a dataset asociate dataesgobr object}
\usage{
load_data(x)
}
\arguments{
\item{x}{dataesgobr containing information and data from datos.gob.es}
}
\value{
A data.frame containing data from datos.gob.es
}
\description{
This function download the data associate to dataset passed like
param and download from datos.gob.es
}
