% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{load_data}
\alias{load_data}
\title{Load data from a file}
\usage{
load_data(file, outfile = NULL)
}
\arguments{
\item{file}{A file with data previously downloaded}

\item{outfile}{A character the path where the file will be save if it need to
be modify, NULL to use the file's directory}
}
\value{
A data.frame
}
\description{
This function loads the data from the file passed like param
}
\examples{
library(dataesgobr)
\donttest{
file <- system.file("extdata", "fichero.csv", package = "dataesgobr")
load_data(file)
load_data(file, outfile = tempdir())
}
}
