% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{load_dataset}
\alias{load_dataset}
\title{Load dataset from a data.frame}
\usage{
load_dataset(dataframe, row = 1)
}
\arguments{
\item{dataframe}{A data.frame with datasets from datos.gob.es}

\item{row}{A number of the row in the data.frame. If the dataset has more
than one element this param determinates the row to load. If row parameter is
missing then the function will load the first row}
}
\value{
A dataesgobr object
}
\description{
Load dataset from a data.frame
}
\examples{
library(dataesgobr)
\donttest{
datasets <- search_by_title('salud')
dataset <- load_dataset(datasets)

datasets <- search_by_title('consumo')
dataset <- load_dataset(datasets, 5)
}
}
